11669: Remove transactions#index and split up TransactionControllable
