+++ app/views/admin/accounting/transactions/_modal.slim
@@ -1,4 +1,3 @@
 .modal.wider id="transaction-modal" tabindex="-1" role="dialog" aria-labelledby="Transaction Modal" data-backdrop="static"
   div.modal-dialog
     div.modal-content
-      = render "admin/accounting/transactions/modal_content"
